Refactor checkCSOAccount for complexity, add userDetails module, refactor stories and tests for duplicated code
