+++ src/frontend/efiling-frontend/src/components/page/home/Home.stories.js
@@ -4,6 +4,7 @@ import { createMemoryHistory } from "history";
 import axios from "axios";
 import MockAdapter from "axios-mock-adapter";
 import { getTestData } from "../../../modules/confirmationPopupTestData";
+import { getUserDetails } from "../../../modules/userDetails";
 
 import Home from "./Home";
 
@@ -33,18 +34,7 @@ const navigation = {
     url: ""
   }
 };
-const userDetails = {
-  bceid: "bobross42",
-  firstName: "Bob",
-  middleName: "Norman",
-  lastName: "Ross",
-  email: "bob.ross@example.com",
-  accounts: [
-    {
-      type: "CSO"
-    }
-  ]
-};
+const userDetails = getUserDetails();
 
 const LoaderStateData = props => {
   mock.onGet(apiRequest).reply(400);
